fix(styled): fixed styles
