+++ src/sentry/static/sentry/app/components/sidebar/sidebarMenuItemLink.tsx
@@ -1,28 +1,33 @@
 import React from 'react';
-import styled from '@emotion/styled';
 
 import Link from 'app/components/links/link';
 import ExternalLink from 'app/components/links/externalLink';
-
-import {OrgSummary} from './sidebarOrgSummary';
 
 export type SidebarMenuItemLinkProps = {
   /**
    * Use this prop if button is a react-router link
    */
-  to: string;
+  to?: string;
   /**
    * Use this prop if button should use a normal (non-react-router) link
    */
-  href: string;
+  href?: string;
   /**
    * Is an external link? (Will open in new tab; Only applicable if `href` is used)
    */
-  external: boolean;
+  external?: boolean;
   /**
    * specifies whether to open the linked document in a new tab
    */
-  openInANewTab: boolean;
+  openInANewTab?: boolean;
+  /**
+   * It is raised when the user clicks on the element - optional
+   */
+  onClick?: () => void;
+  /**
+   * Inline styles
+   */
+  style?: React.CSSProperties;
 };
 
 const SidebarMenuItemLink = ({
@@ -40,7 +45,8 @@ const SidebarMenuItemLink = ({
 
   if (href) {
     return external ? (
-      <ExternalLink {...props} href={href} target={target} />
+      // target is not passed here, as ExternalLink by default opens the link in a new tab
+      <ExternalLink {...props} href={href} />
     ) : (
       <Link href={href} target={target} {...props} />
     );
@@ -49,29 +55,4 @@ const SidebarMenuItemLink = ({
   return <div tabIndex={0} {...props} />;
 };
 
-const StyledSidebarMenuItemLink = styled(SidebarMenuItemLink)`
-  color: ${p => p.theme.gray5};
-  cursor: pointer;
-  display: flex;
-  font-size: 14px;
-  line-height: 32px;
-  padding: 0 ${p => p.theme.sidebar.menuSpacing};
-  position: relative;
-  transition: 0.1s all linear;
-  ${p => (!!p.to || !!p.href) && 'overflow: hidden'};
-
-  &:hover,
-  &:active,
-  &.focus-visible {
-    background: ${p => p.theme.offWhite};
-    color: ${p => p.theme.gray5};
-    outline: none;
-  }
-
-  ${OrgSummary} {
-    padding-left: 0;
-    padding-right: 0;
-  }
-`;
-
-export default StyledSidebarMenuItemLink;
+export default SidebarMenuItemLink;
